PlayerCreatorUtilities has been redesigned
After adding and removing connections to the players, it was noticed that the process of deleting the old player caused problems on slow clients, and the functions were redesigned to not return parameters and return game objects with action when the operations are completed. Thus, after the necessary connection operations are completed, the objects are deleted or such operations are performed.